Fix: cleanup comments and organize
